+++ PeriodicalReservation/PeriodicalReservation/BookingInfo/View/ListWithRadioButtons.swift
@@ -46,18 +46,3 @@ struct ListWithRadioButtons: View {
     }
 }
 
-
-struct RadioButtonView: View {
-    let item: String
-    let isSelected: Bool
-    
-    var body: some View {
-        HStack {
-            Image(systemName: isSelected ? "largecircle.fill.circle" : "circle")
-                .tint(.black)
-            Text(item)
-                .padding(.leading, 5)
-        }
-        .foregroundColor(isSelected ? .blue : .black)
-    }
-}
