+++ PeriodicalReservation/PeriodicalReservation/BookingInfo/ViewModel/RoomsViewModel.swift
@@ -1,0 +1,21 @@
+//
+//  RoomsViewModel.swift
+//  PeriodicalReservation
+//
+//  Created by Ajinkya Adole on 11/12/23.
+//
+
+import Foundation
+
+
+class RoomsViewModel: ObservableObject {
+    
+    @Published var tableData: [String] = ["Ajinkya ", "Ram", "Shiva"]
+    @Published var selectedListItem: String?
+    @Published var isCollapsed: Bool = true
+    
+    init(tableData: [String], isCollapsed: Bool = false) {
+        self.isCollapsed = isCollapsed
+        self.tableData = tableData
+    }
+}
